Improve fallback behavior by picking lowest noise nearest neighbor
In the case when there is a virtual qubit not matched as part of the
partial subgraph matching process we previously just picked an
unallocated physical qubit in index order. This would potentially result
in a really poor layout as the the first unused qubit could be a great
distance away from the subgraph. This commit addresses this issue by
changing the behavior for allocating a qubit outside the subgraph by
finding the nearest neighbors to the already selected physical qubits
and picking the qubit with the lowest 1q error rate from that set.